Cherry-pick MSFT 8690551 for bug check-in to DART branch during October stabilization. PR 99249: Merge dev/miniksa/imelines to dev/main
This turns out to be a result of adding the new TextAttributeRun class to manage the RGB color space.

The refactoring of the code around the Conversion Area inserting its text into the buffer resulted in a small issue with the loop.

It turns out that each time through the loop, a new attribute with the appropriate lines around the conversion area text is supposed to be authored 1-by-1 for each column of the buffer, then inserted into the attribute run.

In reality, the "insert into the run" step was placed outside the loop such that only the attribute for the Nth character would be applied only to the 1st character's position and the rest would be left undecorated.

The correction here is to move the "insert" step back inside the loop.

Verified this restores functionality while typing into the composition line in Japanese IME and restores the left/right cursor position functionality while the composition line is active.

=======

 - Remove unused code that we found while debugging something else.
 - Restore InsertAttrRuns call to inside the loop when accepting text into the
 - Add comment and cleanup leftover lines.

Related work items: #8690551