Cherry-pick MSFT 8426499 for bug check-in to DART branch during October stabilization. PR 98507: It looks like if we stop doing our manual linear compensation during
It looks like if we stop doing our manual linear compensation during WM_DPICHANGED and rely on the new WM_GETDPISCALEDSIZE message that the OS is doing the right thing for us. I'm pretty sure we introduced this compensation code before this message was available to us and are just removing it now.

This improves snapping functionality when crossing a DPI boundary back and forth using a mouse.

I verified that high DPI scenarios still work fine in the following scenarios:
- Dragging window back and forth across 125/150/200% DPI boundary (minimal cursor drift on title bar)
- Snapping across DPI boundaries with the keyboard
- Snapping across DPI boundaries with the mouse
- Attempting cross and cross back across DPI boundary with snap guides with the mouse
- Changing DPI by TS-ing into 100% DPI display from another High DPI device (SP3)
- Changing DPI by TS-ing when full-screened

Everything appears to be working just swimmingly.

Related work items: #8426499