We're going through our PMv1.5 clients and promoting them to PMv2. Here's console!
Console needs to preserve running downlevel, so we're grabbing SetProcessDpiAwarenessContext via GetProcAddr. We also don't want to regress any of the PMv1.5 behavior (at least yet) on downlevel, so I made it call the PMv1.5 dialog scaling API in the fallback case.

One of the big changes with PMv2 is that you don't need to manually opt in HWND trees to the dpi broadcast. So I could've made it so we don't call EnableChildWindowDpiMessage on the main console window if we're PMv2, but it'll add a lot of unnecessary code. As it is, the EnableChildWindowDpiMessage call is a harmless no-op since the broadcast is already enabled at CreateWindow time since the thread has a PMv2 dpi context.

How Verified:
Built and patched conhostv2.dll
Dumped the main console PWND and verified the dpiContext is PMv2 (0x22)
Verified console DPI change behavior on a multi-dpi system:
    Paid special attention to the properties and find dialogs and made sure they scale.
    Played with as many different features as I could, like menu scaling.