PR 97267: Merge dev/migrie/b/8939531-resources-missing to dev/main
When the console moved build systems, the new one didn't have the definitions to build and link the res.rc file. These have been added to the ConhostV2.dll and OpenConsole.exe definitions.

I also added the .ico to an open/res folder, as that was originally loaded from %SDXROOT%\windows\Core\ntuser\server, and the resources wouldn't build without it.

In the future, the resources ideally should be migrated to more up-to-date versions (Visual Studio kept trying to recompile them on me). Right now they're indirectly added as "Included Resources" so that VS won't mangle them, but VS has support for a lot of really nice editors for these kinds of things that we may want to take a look at.

Related work items: #8939531