Cherry-pick MSFT 9185881 for bug check-in to DART branch during October stabilization. PR 107127: Change wcscpy_s to StringCchCopyW to prevent crash/exception being thrown when
 - Change wcscpy_s to StringCchCopyW to prevent crash/exception being thrown when the array fills up.
We don't care if there is an out-of-space error in the telemetry buffer holding these strings. We just want to fill it until full then ignore the rest of the text.

The change to uCRT seems to have activated some handler in the _s versions of the CRT functions that will just kill the app and call Watson. The StrSafe functions don't appear to have this problem.

Related work items: #9185881