PR 161397: Fix for #9358589 and #9842794:
Fix for #9358589 and #9842794:
We're encountering a near-NULL AV encountered in `SrvWriteConsoleInput()` when running input tests. This is due to some validation that was inadvertently removed during the conversion to a separate server library. In particular, `DereferenceIoHandle()` used to return `STATUS_INVALID_HANDLE` in the following cases:
- Handle is `NULL`
- The requesting access mask is `0x0`
- `ulHandleType` was non-zero AND the handle's type didn't match

The latter two cases above (making sure the access mask is acceptable and ensuring the handle types match) are already handled in `ConsoleHandleData::Is[Read|Write][Allowed|Shared]()` and `ConsoleHandleData::Get[Input|Screen]Buffer()`, so the fix is to add a null check on the result of `GetObjectHandle()`. Since we're already in the method, I've also restored the correct behavior of zeroing `a- >NumRecords` and returning `STATUS_INVALID_HANDLE` as we've done in the past.

Related work items: #9842794, #9358589