Merged PR 182142: translate position dependent keycodes to position agnostic
Translates position dependent ctrl/shift keycodes to position agnostic in order to remain consistent with the old behavior. The change in behavior was what broke selection, selection could be modified to also check for the position dependent keycodes but we allow people to inspect the contents of the input buffer so if we didn't change the input record itself then that behavior would also be different.

Related work items: #10142744