Merged PR 207525: Merge dev/migrie/b/9311951-keyup to dev/main
When the v1 code was originally ported, I think this branch was mis-translated. This could cause the keyup for the character to be missed. This resolves this regression from v1, and adds a test to ensure that this sequence of keypresses always generates 4 input events in the queue.

Related work items: #9311951