Merged PR 222173: Merge dev/austdi/fix_alt_gr_test to dev/main
Revises the alt-gr ft test to more accurately test the keypresses when using a German keyboard. Note that because the test needs to change the current keyboard layout to a German one it currently can't be run as part of the CI builds. It can be run from a dev machine by passing "/p:DevTest" to te.exe. Also:

- adds tracing functions for MSG and INPUT_RECORD
- adds custom window messages to change the state of keys and the keyboard layout

Related work items: #10187614