Merged PR 302156: Merge dev/miniksa/title to dev/main
Adds complete set of tests on GetConsoleTitleA/W APIs to ensure compatibility with v1 console behavior.

This regression was introduced as we relayered the server and simultaneously fixed OACR issues inside our code that had marked out buffer overflow conditions.

Our changes had caused the API to return insufficient buffer codes (not always S_OK). For compatibility, we must always return S_OK even when the buffer isn't of sufficient size.

Additionally, our changes had caused the return code from the API to comply with MSDN documentation, but apparently the .NET framework has been reliant on receiving a return code of the buffer size needed (even when the API fails). This has been reverted and we'll follow up to update MSDN.

Lastly, under the A version of the API, giving us a buffer that was exactly the length of the title string (or giving us a zero sized buffer) historically triggered a buffer overflow condition inside the console host heap as the API would null terminate the string even when there wasn't space. I have maintained the fixed behavior (no overflows inside the console host heap) and ensured that the server works with the driver to communicate the lengths/memory copies that are expected to the client application/library.

This fix was tested against:
- emacsclient.exe --version
- The provided C language snippet from Paul to grab the title (from the e-mail thread) and report the length
- A simple Console.Title fetch in a C# console application to ensure the exception isn't thrown (mimicing the reported behavior from the e-mail thread)

Related work items: #11255614