Merged PR 391449: Fix crash in remote.exe
Add ApiNumbers to raw IO requests. The ApiNumber is used by the refactored server platform to distinguish which type of WaitBlock is being called back later. The two Raw IO requests emulate a User-Defined IO request (e.g. ReadFile emulates ReadConsole) but failed to set the ApiNumber. As such, when their WaitBlocks were called back later, the dispatcher got confused over what type of request it was and ran off the rails. An assert was in place to catch this early, but didn't fire on fre. This also wasn't caught earlier because virtually all console applications do not use ReadFile/WriteFile on the console handle anymore but rather use the APIs (the CRT and .NET frameworks also do not use ReadFile/WriteFile).

Filed 11750526 to cover the ReadFile/WriteFile entrypoint specifically with more tests to prevent future regressions.

Related work items: #11739564