Merged PR 382560: DirectX Rendering on OneCoreUAP and Headless Support
**NOTE:** This PR no longer includes the changes to ```ConIoSrv``` since it was moved out of the open source repo. Those changes are now in [this PR](https://microsoft.visualstudio.com/OS/_git/os/pullrequest/409406) instead.

This PR enables DirectX rendering on OneCoreUAP by subsuming the code that formerly constituted ```wddmcon.dll``` into ```ConhostV2``` as yet another renderer. In addition, this PR enables support for headless mode as well as honoring BGFX rendering if requested by the SKU.

Render mode is determined from:

- ```HKLM\System\CurrentControlSet\Control\ConKbd\Headless```
- ```HKLM\System\CurrentControlSet\Control\ConKbd\UseBGFX```
- ```RtlGetDeviceFamilyInfoEnum()```

If ```Headless``` is set and ```0```, then the system is in headless mode: there is no display, and keyboard reading is not initialized. Input API's however remain available (i.e. ```Xlate.c```). Similarly, the OneCore Input Thread quits immediately after recognizing that the system is headless, and performs no further initialization. Input API redirection via the Console IO Server remains available. If the system is not in headless mode, then ```UseBGFX``` is probed. If it is set, its value its honored: if it is ```0```, DirectX is used, otherwise BGFX is used. If neither value is set (i.e. not present in the registry), then the device's family info is probed. If the device family is ```DEVICEFAMILYINFOENUM_UAP```, then DirectX is used. Otherwise, BGFX is used.

All the C code that used to belong to ```wdmmcon.dll``` is now contained in a C++ render engine class, ```WddmConEngine```, ultimately included in ```ConhostV2.dll```. As opposed to relying on the Console IO Server for rendering when using BGFX, each instance of ```conhost``` renders for itself such that no calls into DirectX from ```csrss``` are necessary. However, provided the exclusive nature of the renderer (i.e. there can only be one program controlling the DirectX device), synchronization via the Console IO Server is performed to ensure that no two instances of ```conhost``` attempt to acquire device resources simultaneously. This includes inter-process synchronization during Alt+Tab as well as during console launch. In such instances, the previous owner of the display is asked to relinquish control, and only then may the next client continue.

Related work items: #11636560, #11449211, #11636553