Merged PR 382679: Virtual Terminal Input API Redirection on OneCore
The Virtual Terminal Adapter makes calls into the following Win32 API's that are unavailable on OneCore:

- ```MapVirtualKeyW```
- ```VkKeyScanW```
- ```GetKeyState```

Since there is no input stack on OneCore, their "port" lives in the Console IO Server, which maintains input device state and layout information. In order for the VT decoder to function fully on OneCore, its calls to those API's need to be redirected to the ```ServiceLocator```'s ```LocateInputServices()``` to determine who can service them. When the console is running in full Win32 mode, the calls are responded to by their official implementations, while in OneCore mode, the calls are responded to by the versions in the Console IO Server.

The reason for defining the redirection in a separate header, and their implementation in a separate ```cpp``` file, as opposed to merely putting everything in one header in the ```TerminalAdapter``` project are headers: I couldn't figure out a way to resolve header conflicts (```NTSTATUS``` mainly) other than doing it this way and relying on the linker. It also keeps the VT code clean of extraneous stuff.

Related work items: #11549919, #11449211