Merged PR 475896: GetConsoleAlias API is not returning error codes the way that Clink is expecting
Clink makes some bad assumptions on our API including not checking return codes. This causes it to fail with our newly revised server.

However, in Clink's defense, our API made no sense and doesn't return the correct number of bytes it needs or sensible error codes in most circumstances. Our API also never complied with the documentation given on MSDN.

This change generates a combinatorial test against the GetConsoleAliasA/W API against the v1 console and then reintroduces the broken behaviors to the v2 console server library and API responder code.

This doesn't resolve the non-sense that is this API nor the documentation issue, it just restores the behavior to be compatible with v1 and therefore restores the functionality of the Clink application.

Cherry picked from Pull request 472634.

Related work items: #11524988