Merged PR 487946: Fix incorrect behavior of WriteFile API with UTF-8 characters
When using `WriteFile`, a caller expects to receive the number of bytes consumed out of the stream that it had provided so it knows whether it needs to resubmit part of it later or not.

The UTF-8 parser wasn't really written to return this data in that form. It would instead just return the number of characters that it generated at the end. This means that if invalid characters were in the stream, it would skip them and not generate anything for them, then report back that too few characters were consumed from the stream (sending some callers of `WriteFile` into a potentially infinite loop.)

This combined with an old/bad `*sizeof(wchar_t)` in the server code before the refactor made this work, but it was still incorrect.

After fixing the server layer, it exposed the assumption in the UTF-8 parser that wasn't quite right.

This resolves the issue for this case by teaching the UTF-8 parser the difference between bytes consumed from the original stream and bytes written. This allows the server layer to always properly convey that we "consumed" all the bytes from the original stream even if we had discarded/stripped some of those out for being invalid UTF-8 sequences.

The unit tests for the UTF-8 parser have been updated to accommodate the new function signature as well as test the difference between "consumed" and "generated bytes.

A feature test has been added on top of the `WriteFile` API to submit invalid bytes to our console in UTF-8 mode to ensure that it will report them consumed even when they are discarded.

Cherry picked from Pull request 486461.

Related work items: #12069291