Merged PR 494588: Snapping a command prompt to a corner can result in visual corruption
Mistaken assumption in how we adjusted the invalid rectangle for scrolling operations. It turns out we weren't accounting for the pixels we left behind that we "uncovered" as a result of scrolling what was on top.

I had thought that `ScrollWindowEx` and `ScrollDC` would have had this covered, but it turns out that they do not.

While I was at it, I also fixed the invalid restriction for rectangles to not leave negative pixels in the rectangle. It didn't harm anything (it would be mathed out/excluded later) but it looked weird in the debugger, so I fixed it.

Cherry picked from Pull request 490308.

Related work items: #12053579