Merged PR 499953: Right-sizing window width crazifies background color painting
When traditional resize is on and the buffer width is shrank, we mishandle the resize of the Attributes array that keeps track of colors. It looks like this was just a mistake made when it was adjusted for RGB full color. Simple fix to pass the appropriate length to the function.

Cherry picked from Pull request 499430.

Related work items: #11561936