Merged PR 531746: Improve fitting characters within calculated column sizes when give to GDI.
Add DX parameter to our PolyTextOut structures. This will specify the pixel widths that we want to allow for each individual character in the run so we should get things out of GDI more in line with what we expect based on our internal column estimates. This also updates the IsCharFullWidth function to use A+B+C widths for a more accurate column measurement (aligns it with the font width estimate which is already doing A+B+C).

Used the samples from the linked GitHub threads related to shaded rectangles for testing.

Cherry picked from Pull request 528101.

Related work items: #12177308, #12123951, #11816795