Merged PR 541526: Call NTSTATUS_FROM_HRESULT in ConsoleDispatchRequest
We were using an `HRESULT` as the status value in `ApiSorter::ConsoleDispatchRequest`. This can trigger some undesirable debug spew:

> RTL: RtlNtStatusToDosError(0x80004001): No Valid Win32 Error Mapping
RTL: Edit ntos\rtl\generr.c to correct the problem
RTL: ERROR_MR_MID_NOT_FOUND is being returned

The fix is to convert the `HRESULT` back into an `NTSTATUS`. There was a commented-out call that did this along with a comment saying that we were disabling it due to double printing. This, presumably, was due to macro double-evaluation. Since that's already been fixed, we simply need to resume calling `NTSTATUS_FROM_HRESULT()`. Note that the removed comment referred to a workitem that has already been closed completed.

Cherry picked from Pull request 536404.

Related work items: #12123888