Merged PR 555694: can't type cyrllic text via keyboard to cp-866 windows console applications
Use a local when converting text from Unicode to Codepage on Read/PeekConsoleInputA methods. It turns out that WideCharToMultiByte is sensitive to using the same memory region for source and destination (where as the old Rtl* functions we used earlier were not sensitive to this.) #12294225

I searched the code to find all other usages of this function and confirmed they are not using the same source/destination memory space (they shouldn't be subject to this issue.)

Manually verified versus Far Manager 1.65 (old version from year 2000) that this is resolved with this change when Input & Output CPs are 866 (Russian).
Far Manager 3 (current version) didn't have this issue because it used Read/PeekConsoleInputW methods.

Cherry picked from Pull request 554148.

Related work items: #12294225