Merged PR 568390: UIA classes query screen info and text buffer objects instead of storing
UIA classes query screen info and text buffer objects instead of storing references to them. This fixes a crash that happens when the screen buffer is small and then the window is resized to be bigger than the available rows in the screen buffer, causing new allocations to occur and invalidating the stored references to the text buffer and screen info.

Also fixes using UIA on the alternate screen buffer. Before, UiaTextRange would reference the original screen info even when the alternate one is the active one.

Cherry picked from Pull request 564016.

Related work items: #12347435