Merged PR 582762: Merge dev/paulcam/to-beep-or-not-to-beep to dev/main
Fix for #12383197 - Prior to win32 interactivity being broken out, we would detect if `PlaySound()` was present. If it was, we used it, if it wasn't, we fell back to `Beep()`. At some point during the refactor, we switched to calling `Beep()` when `PlaySound()` failed. This was presumably because the API detection API, namely `IsPlaySoundPresent()`, is private, so we could no longer use it. However, since this code is dependent on win32, and `PlaySound()` is supported on all platforms that support win32, we should just call it and move on with our life.

Related work items: #12383197