Merged PR 652065: [Microsoft/BashOnWindows #2168] Horizontal Lines/Underscores Everywhere
It appears that while using WSL, you can run an application like Vundle (a VIM plugin manager) that generates underscore characters while it is running. If at the time while it is running, you choose to open the property sheet, the underscore character (set as the default fill character in the screen buffer by nature of how VT sequences operate/print) will go into the Property Sheet and be a part of your settings.

Then when the dialog is saved, that underline attribute will be jammed into the LNK file and persisted on future sessions.

There's no way to user-configure the underline setting (or any of the other attributes above the colors portion of the bit flags set).

So I've put a handful of masks at various points in the console code to prevent these flags from leaking out in this way and discarding them in case they show up in unexpected places.

1. In the menu.cpp file, I've put a mask to strip out non-color attributes in the to/from property sheet code to keep flags from leaking this way
2. In the settings startup routines, I've masked out line drawing character flags as a part of attributes (leaving only colors as chooseable) from both LNK and registry file settings.

This resolves the visible issue with the repro scenario given in the GitHub issue. I've tried it with the LNK file that is already broken (that one of the users provided) and it is resolved. I've also tried to reintroduce the issue myself using Vundle and I can no longer do so after this change.

Cherry picked from Pull request 651517.

Related work items: #12319883