Merged PR 695209: Sometimes when changing codepage, the font selected has a different pixel per row count resulting in viewport falling off screen
When updating a font on the screen information, check if we're the active screen information and if there is a window attached to the console. If both are true, ask the window to try to update its physical size to match the rows/columns in the viewport (as the pixel ratio/count may have changed when the font changed.)

Cherry picked from Pull request 694164.

Related work items: #12855493