Merged PR 695274: Merged PR 694642: Merge dev/paulcam/propsheet-dpi to dev/main
Merged PR 694642: Merge dev/paulcam/propsheet-dpi to dev/main

Simple problem. In the propsheet code we generally scale fonts like this:

* Enumerate fonts we want included in the list
* During enumeration, just prior to adding a specific font face/size/boldness to our internal list, we take the font size we enumerated, apply DPI scaling to it, create a new `HFONT` based on that, and stuff it into our data structure

This has the effect of rendering the font at a scaled size even though we are internally accounting for the font size unscaled. The bug here is that at the point where we scale the font during enumeration, we don't have an `HWND` available (needed to determine the scaling factor), but we *do* have an `HDC`. We cheerfully call `WindowFromDC()`, call our DPI functions, and move on with our lives. Unfortunately, `WindowFromDC()` is returning `NULL`, so we end up using the scaling factor of the main monitor (presumably). The fix is to pass the `HWND` of the console.

Related work items: #12881957

Related work items: #12881957