Merged PR 882416: Merged PR 785406: InputBuffer internally stores IInputEvent instead of
Merged PR 785406: InputBuffer internally stores IInputEvent instead of INPUT_RECORD

This is the first step in switching over to an internal representation of our public INPUT_RECORD struct that we can extend to support utf32. InputBuffer still reads INPUT_RECORDs but they are converted and stored as IInputEvent. Writes to InputBuffer are as IInputEvent. Note that for now IInputEvent is basically the same as INPUT_RECORD.

Related work items: #13273221

Related work items: #13273221