Merged PR 1120770: incorrect length given to InsertAttrRuns function =
incorrect length given to InsertAttrRuns function (recently updated to improve performance.) #13505675

I had updated the InsertAttrRuns function to use the inplace algorithm for improved performance and added this assert while I was at it to try to catch bad usages. Turns out there was a bad usage in the IME path. This corrects the values given to the function. (Off by 1 error).

Related work items: #13505675