Merged PR 1134712: Restore cursor invalidation to how it was before VtRenderer
Restore cursor invalidation to how it was before VtRenderer was added in and the abstraction was created to resolve assorted painting issues with the cursor.

Validated with Mike's VtPipeTerm tool that things are still fine for the Vt-based renderer route and validated manually that these issues are resolved when cursor positioning is restored to its previous codepaths.

Cherry picked from !1115996

Related work items: #14495491, #14611160