Merged PR 1116119: Merge dev/migrie/b/14562858-wat to dev/main
There's a lot of weirdness regarding the inheritability of console handles.

This change patches VtPipeTerm to prevent it from hanging on to the conhost handles, and prevent conhost from getting handles it can't reference.

It also makes conhost tear down gracefully when the VtIo pipes are broken, and close the attached client applications.

Related work items: #14562858