Merged PR 1167415: Tweak resize/reflow algorithm to prevent losing a hard newline per feedback bug
When there is text that is inserted into the buffer that is pre-wrapped with a hard newline at the end (e.g. someone printing really long lines with Console.WriteLine() from a C# application), we can end up losing the hard newline at the end of a line when going through the reflow algorithm.

This adjusts the wrap flagging and cursor positioning for the very end of the viewable space of the buffer when the line of text just barely fits into the X dimension of the window.

Also included is a little debug thing that was useful while trying to figure this out to paint a little orange outlined frame around the final character of the buffer IF there was a soft return (wrap was forced and can be unwrapped later, NOT a newline character).

Cherry picked from !1160236

Related work items: #13155961