Merged PR 1198084: Correct byte value returns when WriteConsoleA is serviced off a wait routine
There was an oversight in the server refactoring in March 2017 where we adjusted WriteConsole and its waits to have a better separation between the server portion and the rest of the console code.

We had originally thought that we could simply convert all A methods to W and just call the W versions to do the handling then back-convert the result bytes at the end. And that is true, for calls that don't go into a wait state.

What docker-compose inside VSCode has revealed to us is that WriteConsoleA calls that are delayed momentarily by a wait state (from a selection on the window) can return an incorrect number of "bytes written" to the API call, often more bytes than were originally given.

The solution here is to store some additional contextual data when going into a wait state so we can resynthesize the correct byte count later when the waiting call is dispatched as the selection is released. We have to hold information about the codepage (as it was at the time the wait occurred), some information on whether we stored or retrieved part of a DBCS sequence, and the UTF-8 consumption counter (for CP 65001 only). When the wait is serviced, we notify it of the message's original Unicode/non-Unicode status and it uses this extra state to synthesize the correct byte count for the return packet.

There's a bit of code duplication here between WriteConsoleAImpl and the WriteData::Notify method in terms of this calculation, but eliminating it at this time would require a larger contortion than I'm comfortable with. I tried to author this fix such that it wouldn't be difficult to backport to RS3 should need arise given that this bug was reported originally on the GitHub project Microsoft/VSCode against the Fall Creators Update.

I also threw in a few tests of the ApiRoutines surface into the unit test library to ensure that the various permutations of WriteConsole aren't regressed in this way in the future when we do make a cleaner and more architectural change to WriteConsoleAImpl and WriteData::Notify to eliminate some of the duplication.

Cherry picked from !1194354

Related work items: #14643897