Merged PR 1267085: Merge dev/migrie/r/13994975-refactor-cmdline to dev/main
Pull out the alias and history code into their own files.

I've been meaning to do this for a while, but never got around to doing it the last time I was working on aliases. This will make working in those areas of the codebase, as well as cmdline in general, a lot easier.

13994975 is related, but this does not fix that.

Related work items: #13994975