Merged PR 1303416: Add a canary test for v1 launching
This test will change the console mode to v1 and launch a copy of CMD to ensure that v1 keeps working. It will check to make sure it is "still running" from a process exit code standpoint and redirect ONLY the stdout to try to see the banner that CMD prints out on launch. It doesn't redirect STDIN and asks for a new console session so a completely new window and new conhost/openconsole will open up with the conhostv1.dll loaded.

Also adjusted a bit of the job object control in the Init to keep processes from running away and created a helper to control the registry state of the console v1/v2 selection.

Related work items: #14964735