Merged PR 1346506: Console listens on signal channel for window change message
This change defines the window change message on our signal pipe and applies the change the same way that it would if it had received an in-band window resize message in the VT server input pipe.

For testing, I extended VtPipeTerm and performed a manual test of the functionality.

Related work items: #15536287