Merged PR 1354133: Merge dev/miniksa/vs2017 to dev/main
This changes the project configuration to use the RS3 SDK and the Visual Studio 2017 projects/solutions. This will enable us to move to Package ES Lab E pool which has about double the machines of the Lab D pool and will put us on the most recent Windows SDK and Visual Studio tools.

Related work items: #15608048