Merged PR 1359508: Improve Cursor Positioning in PTY Mode
When moving the cursor due to new text being emitted, we'd invalidate both the location of the text that was emitted, and the new position of the cursor. Almost all of the time, the position that the cursor is getting moved to is a blank space. That would mean that when rendering someone typing "asdf", we'd emit "aSPC^H","sSPC^H","dSPC^H","fSPC^H", rendering both the new character, and the contents of the new cursor position (a space) then also a backspace (to move the cursor back where it is).

Now, we uniquely identify cursor invalidates, so that each engine knows when it's getting a cursor invalidation vs a traditional invalidation. This lets the VT engine not invalidate the cursor position, and only invalidate actual characters.

There's also the problem that if the cursor was outside the "invalid" region, it wouldn't get painted at all. We're removing that limit, so that the VT engine ALWAYS paints the cursor position.

Also, we were using the wrong VT sequence for moving to the end of a Erase Characters section. This could cause tmux to paint it's status line incorrectly, amongst other potential issues.

Cherry picked from !1357915

Related work items: #14711161