Merged PR 1389501: Merge dev/migrie/b/15682977-clear-first-paint to dev/main
This is to better support WSL interop. When the VT renderer first starts, it now emits a ESC[2J and ESC[H, to clear the screen and move the cursor to the origin. This makes it so anything that's already on the screen will be moved into scrollback, rather than painted over.

I've also update the VT renderer to make sure to paint if the cursor was moved. If only the cursor moves in a frame (but no text changes), we wouldn't have anything invalid, and we wouldn't emit the new cursor position. This stops us from skipping those paints, and instead now we will emit the new cursor position.

Cherry picked from !1379788

Related work items: #15682977