Merged PR 1416735: Merge dev/migrie/b/15813316-inherit-cursor to dev/main
This is the bulk of the work remaining to get interop via conpty working.

Primarily, this adds a new argument to conhost, `--inheritcursor`. When a terminal uses that arg with conhost, conhost will emit a VT sequence on startup to try and get the current cursor position of the end terminal. The terminal (for ex, tmux, or conhost if there's nothing else running as a pty master) will respond with the position of the cursor on the input pipe, which conhost will read, and then move the cursor to that position. Conhost will then continue on it's merry way of starting up.

All of this happens *before* the client application is started, so client's output will start in the inherited location in the conpty.

We'll also *not* paint the region of the buffer above the inherited cursor position, until the contents of the conpty naturally get there. This is to prevent conpty from painting the empty region of the buffer to the terminal.

See also MSFT: 15823972 and [OS PR:1412191](https://microsoft.visualstudio.com/_git/os/pullrequest/1412191) for the WSL half of this change.

CC @<benhill@ntdev.microsoft.com> @<jostarks@ntdev.microsoft.com>

Once this PR is in, it'll be checked into ..._ACI_OSS, then cherry-picked to the feature/low_rs4 branch for staging.

Cherry picked from !1412298

Related work items: #15813316