Merged PR 1585111: Merged PR 1578523: Merge dev/migrie/f/intercept-C-S-cv to dev/main
Merged PR 1578523: Merge dev/migrie/f/intercept-C-S-cv to dev/main

Adds a setting to override all input modes to intercept C-S-v as a paste keystroke.

When the setting is true, C-S-v will paste, even in raw input mode (yes, in wsl and PSReadline)

C-S-c was also added as a copy shortcut, for symmetry. Copying was never the problem, but it would be weird for only one to work.

Related work items: #16460490

Related work items: #16460490