Merged PR 1619080: Fix a crash when running WSL interop
If you close the console terminal window while running cmd.exe in wsl, we'll break into the debugger. This is because we'll try and paint one last time before exiting (due to `TriggerTeardown`), however, we're tearing down because the other side of the pipes closed the pipes. So when we would paint, there would be nothing on the other side of the pipe, and WriteFile would fail. Most paint steps only log on errors, but StartPaint would throw, and the Vt Renderer can attempt to write during that phase.

So, the WriteFile would fail, causing StartPaint to fail, causing PaintEngine to throw, which would bubble through the Teardown handler, causing a crash instead of a graceful shutdown.

The easy solution is to just log on a failure in StartPaint _IF_ we've started tearing down.

There was a related issue where sometimes the input thread would crash when closing interop, but that was because of this as well.

Cherry picked from !1595271

Related work items: #16503902