Merged PR 1638418: Change aliases to use modern structures/algorithm
This fixes the aliases.cpp file to use a single STL collection to store data over our custom DDK-based list structures.
It also updates the match and copy algorithm to use STL and WIL for assembling the final string.
This also eliminates some analysis exceptions and the manipulation of raw pointers moving to all bounds-checked operations.
It includes a comprehensive unit test for the match and copy algorithm based on the previous edition of the code before I made the changes.

Related work items: #8805227