Merged PR 1678718: Merged PR 1626145: Merge dev/migrie/f/15933349-fallthrough-sequences to
Merged PR 1626145: Merge dev/migrie/f/15933349-fallthrough-sequences to dev/main

There are some VT sequences that are really only helpful to the end terminal, for example, setting the values of the color table, setting the cursor attributes (not the cursor position). For these sequences, instead of having us consume the sequences and then re-render them back to the tty, we should instead just let them pass-through to the tty.

Now, when the adapter realizes it's a pty and a sequence is meant for the tty, we can throw a TerminalSequenceException. This will get caught by the StateMachine (who still knows the contents of the string), and the StateMachine will write that string straight to the renderer.  This is enabled by adding a ActionPassThroughString to each StateMachineEngine, which does nothing outside of pty mode, does nothing different then PrintString for the input engine, and writes the string to the renderer for the output engine in pty mode.

Related work items: #15933349, #16486043

Related work items: #15933349, #16486043