Merged PR 1706178: Move output buffer code into its own library and enforce modern standards
This moves the output buffer code into its own library and sets the Audit Mode on it to enforce Modern CPP standards on it.

I have fixed several hundred warnings related to manipulating buffers, excepts, consts, etc. in the code while moving it.
I've also updated pretty much all of AttrRow (and the related tests and callers) to deal entirely in std::vectors (which eliminated hundreds of errors).

The remaining warnings need consultation with Austin to determine what to do with some of these classes especially as the buffer design is changing again. They're not warnings-as-errors right now, so I'll check it in like this to get it broken apart into a lib and *mostly* fixed and we can fix the remaining issues when Austin returns.

Related work items: #16479368